move generator-base tests to generators/base
